requirements.txt
beautifulsoup4==4.7.1
bs4==0.0.1
cachetools==3.1.1
certifi==2019.3.9
chardet==3.0.4
Click==7.0
fuzzywuzzy==0.17.0
google-api-python-client==1.7.9
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.3.0
gTTS==2.0.3
gTTS-token==1.1.3
httplib2==0.12.3
idna==2.8
oauthlib==3.0.1
peewee==3.9.5
pyasn1==0.4.5
pyasn1-modules==0.2.5
PyAudio==0.2.11
pycairo==1.18.1
pydbus==0.6.0
pydub==0.23.1
PyGObject==3.32.1
PyGObject-stubs==0.0.2
python-Levenshtein==0.12.0
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
six==1.12.0
soupsieve==1.9.1
SpeechRecognition==3.8.1
uritemplate==3.0.0
urllib3==1.25.3
